# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Currency exchange rate API
  version: 1.0.0
paths:
  /api/v1/currencies:
    get:
      summary: Get list of all available currencies
      operationId: getCurrencies
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                minItems: 0
                items:
                  type : string
        '404':
          description: not found
        '500':
          description: Unhandled error by server

  /api/v1/conversion-rates:
    get:
      summary: Get list of all available conversion rates
      operationId: getConversionRates
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                minItems: 0
                items:
                  $ref: '#/components/schemas/CurrencyConversionRateResponse'
        '404':
          description: not found
        '500':
          description: Unhandled error by server

  /api/v1/conversion-rates/date/{date}:
    get:
      summary: Get list of all available conversion rates for given date.
      operationId: getConversionRatesByDate
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                minItems: 0
                items:
                  $ref: '#/components/schemas/CurrencyConversionRateResponse'
        '404':
          description: not found
        '500':
          description: Unhandled error by server

  /api/v1/conversion-rates/currency/{currencyCode}/date/{date}:
    get:
      summary: Get list of all available conversion rates for given date.
      operationId: getConversionRatesByCurrencyAndDate
      parameters:
        - name: currencyCode
          in: path
          required: true
          schema:
            type: string
            example: "USD"
        - name: date
          in: path
          required: true
          schema:
            type: string
            example: "2021-01-22"
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrencyConversionRateResponse'
        '404':
          description: not found
        '500':
          description: Unhandled error by server

components:
  schemas:
    CurrencyConversionRateResponse:
      type: object
      description: contains currency exchange rate details
      properties:
        currencyCode:
          type: string
          example: "GBP"
        conversionRate:
          type: double
          example: 12.5
        rateDate:
          type: string
          format: date
          example: "2025-10-25"